requirements.txt
Flask
gevent
numpy
Pillow
tensorflow-cpu
Werkzeug
Gunicorn
protobuf==3.20.*